Project splitting. Infrastructure (Services and Interfaces from WebStore were moved to folder Services to new Projects. The links between projects were  made.
