+++ Common/WebStore.DomainNew/ViewModels/CartViewModel.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 
-namespace WebStore.ViewModels
+namespace WebStore.DomainNew.ViewModels
 {
     public class CartViewModel
     {
