+++ Common/WebStore.DomainNew/ViewModels/RegisterUserViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace WebStore.ViewModels
+namespace WebStore.DomainNew.ViewModels
 {
     public class RegisterUserViewModel
     {
