+++ Services/WebStore.Interfaces/IOrderService.cs
@@ -1,8 +1,8 @@
 ﻿using System.Collections.Generic;
 using WebStore.DomainNew.Entities;
-using WebStore.ViewModels;
+using WebStore.DomainNew.ViewModels;
 
-namespace WebStore.Infrasructure.Interfaces
+namespace WebStore.Interfaces
 {
     public interface IOrderService
     {
