+++ Ui/WebStore/Views/Account/Login.cshtml
@@ -1,4 +1,4 @@
-﻿@model LoginViewModel
+﻿@model WebStore.DomainNew.ViewModels.LoginViewModel
 @{
     ViewBag.Title = "Логин";
 }
