+++ Ui/WebStore/Views/Account/Register.cshtml
@@ -1,4 +1,4 @@
-﻿@model RegisterUserViewModel
+﻿@model WebStore.DomainNew.ViewModels.RegisterUserViewModel
 @{
     ViewBag.Title = "Регистрация";
 }
