+++ Ui/WebStore/Views/Cart/Details.cshtml
@@ -1,4 +1,4 @@
-﻿@model OrderDetailsViewModel
+﻿@model WebStore.DomainNew.ViewModels.OrderDetailsViewModel
 @{
     ViewBag.Title = "Cart";
 }
