+++ Ui/WebStore/Views/Cart/Partial/_ChooseArea.cshtml
@@ -1,4 +1,4 @@
-﻿@model OrderViewModel
+﻿@model WebStore.DomainNew.ViewModels.OrderViewModel
 
 <div>
     <input asp-for="Name" placeholder="@nameof(Model.Name)" />
