+++ Ui/WebStore/Views/Cart/Partial/_OrderPartial.cshtml
@@ -1,4 +1,4 @@
-﻿@model OrderViewModel
+﻿@model WebStore.DomainNew.ViewModels.OrderViewModel
 
 <form method="post" asp-antiforgery="true" asp-action="CheckOut">
     <div class="col-sm-6">
