+++ Ui/WebStore/Views/Catalog/Partial/_ProductDetails.cshtml
@@ -1,4 +1,4 @@
-﻿@model ProductViewModel
+﻿@model WebStore.DomainNew.ViewModels.ProductViewModel
 
 <div class="product-details">
     <!--product-details-->
