+++ Ui/WebStore/Views/Catalog/Partial/_ProductItems.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<ProductViewModel>;
+﻿@model IEnumerable<WebStore.DomainNew.ViewModels.ProductViewModel>;
 
 <!--features_items-->
 <h2 class="title text-center">Features Items</h2>
