+++ Ui/WebStore/Views/Catalog/ProductDetails.cshtml
@@ -1,4 +1,4 @@
-﻿@model ProductViewModel
+﻿@model WebStore.DomainNew.ViewModels.ProductViewModel
 @{
     ViewBag.Title = "Product details";
 }
