+++ Ui/WebStore/Views/Catalog/Shop.cshtml
@@ -1,4 +1,4 @@
-﻿@model CatalogViewModel;
+﻿@model WebStore.DomainNew.ViewModels.CatalogViewModel;
 @{
     ViewBag.Title = "Shop";
 }
