+++ Ui/WebStore/Views/Employee/Details.cshtml
@@ -1,4 +1,4 @@
-﻿@model EmployeeView
+﻿@model WebStore.DomainNew.ViewModels.EmployeeView
 <div class="container">
     @*Информация по выбранному сотруднику с дополнительными свойствами*@
     <div class="col-md-4">
