+++ Ui/WebStore/Views/Employee/Edit.cshtml
@@ -1,4 +1,4 @@
-﻿@model EmployeeView;
+﻿@model WebStore.DomainNew.ViewModels.EmployeeView;
 @{
     ViewBag.Title = "Редактирование сотрудника";
 }
