+++ Ui/WebStore/Views/Employee/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<EmployeeView>
+﻿@model IEnumerable<WebStore.DomainNew.ViewModels.EmployeeView>
 
 @{
     var isAdmin = User.IsInRole("Administrator");
