+++ Ui/WebStore/Views/Profile/Orders.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<UserOrderViewModel>
+﻿@model IEnumerable<WebStore.DomainNew.ViewModels.UserOrderViewModel>
 @{
     Layout = "_ProfileLayout";
     ViewBag.Title = "Мои заказы";
