+++ Ui/WebStore/Views/Shared/Components/Brands/Default.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<BrandViewModel>;
+﻿@model IEnumerable<WebStore.DomainNew.ViewModels.BrandViewModel>;
 
 <div class="brands_products">
     <!--brands_products-->
