+++ Ui/WebStore/Views/Shared/Components/Categories/Default.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<CategoryViewModel>;
+﻿@model IEnumerable<WebStore.DomainNew.ViewModels.CategoryViewModel>;
 
 <h2>Category</h2>
 <div class="panel-group category-products" id="accordian">
