+++ Ui/WebStore/Views/Shared/Partial/_ReviewAndPayment.cshtml
@@ -1,4 +1,4 @@
-﻿@model CartViewModel
+﻿@model WebStore.DomainNew.ViewModels.CartViewModel
 
 @foreach (var item in Model.Items)
 {
