+++ Ui/WebStore/Views/_ViewImports.cshtml
@@ -1,5 +1,4 @@
 ﻿@*Пространство имен моделей*@
-@using WebStore.ViewModels;
 
 @*Тег-хелперы*@
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
