+++ WebStore.sln
@@ -15,6 +15,10 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebStore.DAL", "Services\We
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebStore", "Ui\WebStore\WebStore.csproj", "{7EBD27B6-D2A8-4CC2-B405-54703BC54DB8}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebStore.Services", "Services\WebStore.Services\WebStore.Services.csproj", "{235D4760-B2D6-43F1-B5B6-585A20AAAF1E}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebStore.Interfaces", "Services\WebStore.Interfaces\WebStore.Interfaces.csproj", "{9EA70DCC-B4A9-4453-8CB9-D6B5683FAE6C}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -33,6 +37,14 @@ Global
 		{7EBD27B6-D2A8-4CC2-B405-54703BC54DB8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7EBD27B6-D2A8-4CC2-B405-54703BC54DB8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7EBD27B6-D2A8-4CC2-B405-54703BC54DB8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{235D4760-B2D6-43F1-B5B6-585A20AAAF1E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{235D4760-B2D6-43F1-B5B6-585A20AAAF1E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{235D4760-B2D6-43F1-B5B6-585A20AAAF1E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{235D4760-B2D6-43F1-B5B6-585A20AAAF1E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9EA70DCC-B4A9-4453-8CB9-D6B5683FAE6C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9EA70DCC-B4A9-4453-8CB9-D6B5683FAE6C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9EA70DCC-B4A9-4453-8CB9-D6B5683FAE6C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9EA70DCC-B4A9-4453-8CB9-D6B5683FAE6C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -41,6 +53,8 @@ Global
 		{6C857CC2-471F-44C8-914F-E06B0072B0CA} = {5E65E749-3A05-4E39-B7BF-1A6F00CED9E3}
 		{B0FF40C9-F222-494A-826A-12BA1B1B9548} = {9F894CD8-A9B8-4DD2-A141-66F0F8888A87}
 		{7EBD27B6-D2A8-4CC2-B405-54703BC54DB8} = {0985310C-E0FB-46C5-AEC4-B0300CB3CE2F}
+		{235D4760-B2D6-43F1-B5B6-585A20AAAF1E} = {9F894CD8-A9B8-4DD2-A141-66F0F8888A87}
+		{9EA70DCC-B4A9-4453-8CB9-D6B5683FAE6C} = {9F894CD8-A9B8-4DD2-A141-66F0F8888A87}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {DABD886B-2B4A-419E-A808-AC7A41F56068}
